rename utils file & tests
