--- dev/main.ts
@@ -1,5 +1,0 @@
-import { Unit } from './compiled';
-
-const newUnit = Unit.fake();
-
-console.log(newUnit);
