--- dev/types/file-one.ts
@@ -1,28 +1,0 @@
-export interface Unit {
-    name: string;
-    isLanded?: boolean;
-    address: {
-        line1: string;
-        line2: string;
-    }
-    // health: number;
-    // cost: [number, string, boolean];
-    // blah: number[];
-    // blah2: Array<{age: number, name: string}>;
-    // translations: {
-    //     key1: string;
-    //     key2: string;
-    //     foo: {
-    //         bar: {
-    //             age: number;
-    //         }
-    //     }
-    // };
-    // isLanded?: boolean;
-    // person: Person;
-}
-
-// interface Person {
-//     name: string;
-//     age: number;
-// }
