+++ src/shared/types.d.ts
@@ -1,4 +1,12 @@
+import { tsTypes } from './enums';
+
 export interface options<T> {
     fake?: boolean;
     value?: T;
+}
+
+export interface RawType {
+	key: string | null;
+	type: tsTypes;
+	children: RawType[] | undefined;
 }
