+++ test/compiler/generators/fake.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { recursivlyTraverse } from '../../../src/utils/newLooper';
+import { recursivlyTraverse } from '../../../src/utils/utils';
 import { generateFakeMethod } from '../../../src/compiler/generators/fake';
 
 import { testProject } from '../../utils';
